requirements.txt
psycopg2-binary
requests
yfinance
beautifulsoup4
pandas
sqlalchemy
lxml
python-dotenv
